push code

+++ src/main/kotlin/com/base/onlinelib/Repositories.kt
@@ -1,11 +1,10 @@
 package com.base.onlinelib
 
 import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AuthorRepository : JpaRepository<Author, Int>
+interface AuthorRepository : JpaRepository<Author, Long>
 
 @Repository
-interface BookRepository : JpaRepository<Book, Int>
+interface BookRepository : JpaRepository<Book, Long>
